refactor: Update JWT secret to use environment variable
The code changes in the `jwt.js` file update the JWT secret to use the `process.env.JWT_SECRET` environment variable. This allows for better security and flexibility in different environments. If the environment variable is not set, it falls back to the default value of 'rahasia'. This change improves the security and configuration management of the application.