Update configuration to use environment variables (#10)
* feat: Update app.js to use environment variables for configuration

The code changes in the `app.js` file update the configuration to use environment variables for the `NODE_ENV` and `PORT` values. This allows for better flexibility and configuration management in different environments. Additionally, the `cors` middleware is added to enable Cross-Origin Resource Sharing. This change improves the configuration and security of the application.

This commit message follows the established convention of starting with a type (`feat`) followed by a concise and descriptive summary of the changes.

* feat: Add cors and dotenv npm dependencies

The code changes in the `package-lock.json` file add the `cors` and `dotenv` npm dependencies. This allows for enabling Cross-Origin Resource Sharing (CORS) and loading environment variables from a `.env` file. These changes improve the functionality and security of the application.

This commit message follows the established convention of starting with a type (`feat`) followed by a concise and descriptive summary of the changes.

* refactor: Update MongoDB connection to use environment variables

The code changes in the `mongodb.js` file update the MongoDB connection to use environment variables for the MongoDB URI and database name. This allows for better flexibility and configuration management in different environments. The `process.env.MONGODB_URI` and `process.env.MONGODB_DBNAME` variables are used, with fallback values for local development. This change improves the configuration and security of the application.

This commit message follows the established convention of starting with a type (`refactor`) followed by a concise and descriptive summary of the changes.

* refactor: Update JWT secret to use environment variable

The code changes in the `jwt.js` file update the JWT secret to use the `process.env.JWT_SECRET` environment variable. This allows for better security and flexibility in different environments. If the environment variable is not set, it falls back to the default value of 'rahasia'. This change improves the security and configuration management of the application.