Get connection (#15)
* feat: Add route for connections in index.js

The code changes in the `index.js` file add a new route for `/connections` in the Express router. This allows for handling requests related to connections. This change improves the functionality of the application.

This commit message follows the established convention of starting with a type (`feat`) followed by a concise and descriptive summary of the changes.

* feat: Add route for connections in connection.js

Add a new route for `/connections` in the Express router. This allows for handling requests related to connections. This change improves the functionality of the application.

* feat: Add getAllConnections method to ConnectionController

This commit adds a new method `getAllConnections` to the `ConnectionController` class in the `connection.controller.js` file. The method retrieves all connections for a user and sends a JSON response with the connections. This change improves the functionality of the application.

* feat: Add Connection model and findAllConnection method

This commit adds the Connection model and the findAllConnection method to retrieve all connections for a user. The method uses MongoDB aggregation to match the user's gender and lookup the corresponding user information. This change improves the functionality of the application.

This commit message follows the established convention of starting with a type (`feat`) followed by a concise and descriptive summary of the changes.

* docs: update docs

* feat: Add config file to gitignore (#14)

This commit adds the `config` file to the `.gitignore` to exclude it from version control. This change improves the repository's organization and prevents sensitive configuration information from being committed.

This commit message follows the established convention of starting with a type (`feat`) followed by a concise and descriptive summary of the changes.