Fix preferece (#21)
* refactor: Update Jest configuration to use environment variables

The code changes in the `jest.config.js` file update the Jest configuration to use environment variables for various properties such as `coverageDirectory`, `coverageProvider`, and `preset`. This allows for better flexibility and configuration management in different environments. By using environment variables, the configuration can be easily customized without modifying the code. This change improves the configuration and maintainability of the test suite.

This commit message follows the established convention of starting with a type (`refactor`) followed by a concise and descriptive summary of the changes.

* refactor: Update index route to return JSON response

The code changes in the `index.js` file update the root route to return a JSON response instead of a plain text response. This change improves the API's response format and follows best practices for RESTful APIs.

This commit message follows the established convention of starting with a type (`refactor`) followed by a concise and descriptive summary of the changes.

* refactor: Update MongoDB connection to use environment variables

The code changes in the `mongodb.js` file update the MongoDB connection to use environment variables for the MongoDB URI and database name. This change improves the configuration and security of the application.

* refactor: Update devDependencies in package.json

Update devDependencies in package.json to include "@shelf/jest-mongodb": "^4.3.2", "jest": "^29.7.0", and "supertest": "^7.0.0". This change improves the test suite by adding necessary dependencies for testing MongoDB and making HTTP requests.

This commit message follows the established convention of starting with a type (`refactor`) followed by a concise and descriptive summary of the changes.

* feat: Add test data for connections and swipes

This commit adds test data for connections and swipes in the `test.connections.json` and `test.swipes.json` files. The data includes sample connections and swipes between users. This change improves the testing capabilities of the application by providing realistic data for testing the connection and swipe functionality.

Note: Please remove any unnecessary details from the commit message before using it.

* feat; add test

* refactor: Update user schema preference validation

The code changes in the `user.schema.js` file update the `preference` field validation in the user schema. The previous implementation used `z.array(z.string())`, but it has been changed to `z.array({ message: "Preference is required" })` to provide a more descriptive error message. This change improves the clarity and usability of the validation for the `preference` field.

Note: Please remove any unnecessary details from the commit message before using it.

* refactor: Update MongoDB connection to use environment variables