feat: add validation for preference and add testing for regsiter and login (#23)
* refactor: Update user schema preference validation

The code changes in the `user.schema.js` file update the `preference` field validation in the user schema. The previous implementation used `z.array(z.string())`, but it has been changed to `z.array({ message: "Preference is required" })` to provide a more descriptive error message. This change improves the clarity and usability of the validation for the `preference` field.

* refactor: refactor jhon doe preference

* feat: add testing for POST /user/register and POST /user/login