Add Swipe functionality and route (#25)
* feat: Add SwipeController and Swipe model

The code changes add a new SwipeController class and Swipe model to handle swiping functionality. The SwipeController class includes methods for retrieving all swipes and swiping a user. The Swipe model includes methods for finding all swipes and adding a new swipe. This change improves the functionality of the application by providing the necessary components for managing swipes.

* feat: Add route for /swipe in index.js

The code changes in the `index.js` file add a new route for `/swipe` in the Express router. This allows for handling requests related to swiping functionality. This change improves the functionality of the application by providing the necessary route for managing swipes.

* refactor: Remove unused swipeUser method from UserController

The code changes remove the unused swipeUser method from the UserController class in the user.controller.js file. This method was not being used and can be safely removed to improve code cleanliness and maintainability.

* refactor: errorhanlder, commenting unused default response