feat: add testing and fix some bugs (#27)
* refactor: Remove console.log statements in user and swipe controllers

* feat: Add swipeStatus filter to Swipe model query

The code changes in the `swipe.model.js` file add a `swipeStatus` filter to the query in the `Swipe` model. This filter allows retrieving swipes with a specific status, in this case, "accepted". This change improves the functionality of the application by providing the ability to filter swipes based on their status.

Note: Please remove any unnecessary details from the commit message before using it.